changes variable to const and moves event type to its own file
